patched vulnerability #1
CodeSQL was complaining about the secret key being printed to stdout.
Yes I could have just removed the print statement, but I did want some
kind of cool confirmation that the key was there and all right, so I
generated a hash with the SHA256 alg and printed that. probably best I
have the hashlib module imported anyways, I'll have to see if I can use
that for something later.